Implement pluggable logger
This commit adds `Logger` to the `ServerOpts` so it can be specified by
a library user. `Logger` is now an interface which defines the required
methods a logger must implement.
The default logger implementation, `StdLogger` which is used when no
logger was specified in the `ServerOpts`, is stateless, thus the
`sessionID` parameters is now given as argument to each `logger.Print*`
call.