Call Driver.CheckPasswd if present instead of ServerOpts.Auth.CheckPasswd - Fixes #117 (#119)
Call Driver.CheckPasswd if present instead of ServerOpts.Auth.CheckPasswd - Fixes #117

This allows Drivers to keep track of the Authed user which wasn't
possible before.

This is a backwards compatible change as it adds an optional interface
to the Driver interface which will be called if present and ignored
otherwise.

Co-authored-by: Nick Craig-Wood <nick@craig-wood.com>
Reviewed-on: https://gitea.com/goftp/server/pulls/119
Reviewed-by: Lunny Xiao <xiaolunwen@gmail.com>