Factor into server and driver directories (#120)
Factor into core and driver directories

The change factors the pure server implementation into the "core"
directory and factors the file driver into "driver/file" and the minio
driver into "driver/minio".

This means that users of this library can import goftp.io/server/core
without having to import the file and minio drivers which may not be
needed.

This also adds a compatibility layer which exports all the types,
functions and variables that were exported at the top level.

This means this change should be 100% backwards compatible.

Fixes #116

Co-authored-by: Nick Craig-Wood <nick@craig-wood.com>
Reviewed-on: https://gitea.com/goftp/server/pulls/120